Use ack-grep line number and file name color code for font lock
Use the ack-grep line number and file name color code directly for font
lock. This is to address a problem that file name font lock wouldn't
work for a file name starts with numbers and dash. For example, if a
file name was "1234-5", it was recognized as a line number.